AI-Edu-recomend/backend/requirements-old.txt
# fastapi
# uvicorn[standard]
# pymongo
# requests
# python-dotenv
# beautifulsoup4
# pydantic

# --- Core Server ---
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.7.4

# --- Data & AI ---
sentence-transformers==2.7.0
torch==2.3.1
transformers==4.41.2
# openai==1.30.1
google-generativeai==0.7.2


# --- Database (if using MongoDB) ---
pymongo==4.7.3
motor==3.5.1

# --- Web Scraping (for fetching resources) ---
requests==2.31.0
beautifulsoup4==4.12.3
google-api-python-client==2.137.0

# --- Utilities ---
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.4

lxml

